requirements.txt
Django==3.2.4
djangorestframework==3.12.4
djoser==2.1.0
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
django-filter==2.4.0
django-environ==0.4.5
drf-yasg==1.20.0
django-cors-headers==3.8.0
postgres==3.0.0
# install incase you receive ImportError: cannot import name 'six' from 'django.utils' error
#django-utils-six==2.0
djangorestframework-simplejwt==4.7.1
djangorestframework-jwt==1.11.0
django-currentuser==0.5.3
asyncio-nats-streaming==0.4.0